Enable building the sopv subset
in sop 0.7.2 it's relatively simple to build the sopv subset, for a
reduced-size binary that might be used in compact verification-only
contexts like OS installer images.

This change means that you can do:

  cargo build --no-default-features --features=cliv

and the generated rsop will only have `version`, `verify`, and
`inline-verify` subcommands.

It would be nice to also change the name of the generated binary to
rsopv in this case, but my rust skills are too weak to know how to do
that (though i know how to `mv rsop rsopv` after build ☺)